Make gopack installable via go install
Previously if you tried to `go install github.com/cloudsmith-io/gopack`
you would get a message about the module not having a main package.

This commit rearranges the module so that the top-level package is a
main package, while still keeping the implementation details of the
cobra command separated out from the archiving & modfile-parsing
packages. This means that `go install` can now install the `gopack`
binary.